requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages

#apache-airflow-providers-amazon
#apache-airflow-providers-github
#apache-airflow-providers-slack
#apache-airflow-providers-snowflake
#apache-airflow-providers-tableau
#airflow-provider-fivetran-async
#boto3
#flake8-black
#pendulum
#snowflake-connector-python
#xmltodict
tableauserverclient
python-dotenv